feat: refactor components Mentions: #70
